Entry color support, local storage and bug fix

+++ Balto.UI/Balto/src/app/core/services/local-storage.service.ts
@@ -9,6 +9,11 @@ export class LocalStorageService {
   constructor() { }
 
   public get(key: string): any {
+    if(!this.checkService()) {
+      console.error('Local storage service is unavailable!');
+      return null;
+    }
+
     const itemSerialized = localStorage.getItem(key);
     if(itemSerialized !== null) {
       const item = JSON.parse(itemSerialized);
@@ -24,6 +29,11 @@ export class LocalStorageService {
   }
 
   public set(options: LocalStorageOptions): void {
+    if(!this.checkService()) {
+      console.error('Local storage service is unavailable!');
+      return;
+    }
+
     options.expirationMinutes = options.expirationMinutes || 0;
     const expirationMilisec = (options.expirationMinutes !== 0) ? options.expirationMinutes * 60 * 1000 : 0;
 
@@ -37,10 +47,37 @@ export class LocalStorageService {
   }
 
   public unset(key: string): void {
+    if(!this.checkService()) {
+      console.error('Local storage service is unavailable!');
+      return;
+    }
+
     localStorage.removeItem(key);
   }
 
   public drop(): void {
+    if(!this.checkService()) {
+      console.error('Local storage service is unavailable!');
+      return;
+    }
+
     localStorage.clear();
+  }
+
+  private checkService(): boolean {
+    const testValue = 'LOCAL_STORAGE_TEST';
+    try {
+      localStorage.setItem(testValue, testValue);
+      localStorage.removeItem(testValue);
+      return true;
+    } catch(e) {
+      return false;
+    }
+  }
+
+  public test(): boolean {
+    if(this.checkService()) return true;
+    console.error('Local storage service is unavailable!');
+    return false;
   }
 }
